Warning ritkítás & Kommentek & BugFix
Főképp size_t okozta warningok ritkítása.
(kérdéses, hogy mi történik, ha nincs egyetlen negyedéve sem)
Ha jól emlékszem, akkor nem létezhet a részvény, az alapján ellenőrzi a részvény valódiságát.

Adat frissítő menüt visszaolvasva nem volt egyértelmű rögtön néhány dolog.
Kikommenteztem fontosabb részeit.

Másolásos hibát vétettem a napi nézet gombjainak beviteli kritériumába, melyt javítottam.

+++ projekt/GorbeKategorizalas/main2.cpp
@@ -111,7 +111,6 @@ void main2( SDL_Window &window, SDL_Renderer &renderer){
     cout<<"zz"<<endl;
     loadStock3("TSLA",stock3);
     cout<<"zzz"<<endl;
-    /*
     loadStock2("MCD",stock4);
     cout<<"zzzz"<<endl;
     loadStock2("WMT",stock5);
@@ -133,7 +132,7 @@ void main2( SDL_Window &window, SDL_Renderer &renderer){
     vector<string> reszvenyekNeve = osszesReszveny();
     vector<thread> szalak; szalak.resize(32);
     vector<Stock> stocks; stocks.resize(32);
-    for (int i=1000; i<reszvenyekNeve.size(); i++){
+    for (size_t i=1000; i<reszvenyekNeve.size(); i++){
         for (int j=0; j<32; j++){
             szalak[j] = thread(loadStock,reszvenyekNeve[i],ref(stocks[j]));
             i++;
